LPS-55068 Move
