+++ modules/util/portal-tools-wsdd-builder/src/com/liferay/portal/tools/wsdd/builder/WSDDMerger.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.tools;
+package com.liferay.portal.tools.wsdd.merger;
 
 import com.liferay.portal.kernel.util.FileUtil;
 import com.liferay.portal.kernel.util.StringUtil;
