+++ modules/util/portal-tools-wsdd-builder/src/com/liferay/portal/tools/wsdd/builder/dependencies/server-config.wsdd
